+++ src/Models/Bank.php
@@ -10,10 +10,6 @@ namespace AndrewSvirin\Ebics\Models;
  */
 final class Bank
 {
-    public const VERSION_24 = 'VERSION_24';
-    public const VERSION_25 = 'VERSION_25';
-    public const VERSION_30 = 'VERSION_30';
-
     /**
      * The HostID of the bank.
      */
@@ -44,22 +40,15 @@ final class Bank
     private ?string $serverName = null;
 
     /**
-     * The Server Version of the bank.
-     */
-    private string $version;
-
-    /**
      * Constructor.
      *
      * @param string $hostId
      * @param string $url
-     * @param string $version
      */
-    public function __construct(string $hostId, string $url, string $version = self::VERSION_25)
+    public function __construct(string $hostId, string $url)
     {
         $this->hostId = $hostId;
         $this->url = $url;
-        $this->version = $version;
     }
 
     /**
@@ -130,13 +119,5 @@ final class Bank
     public function getServerName(): ?string
     {
         return $this->serverName;
-    }
-
-    /**
-     * @return string|null
-     */
-    public function getVersion(): ?string
-    {
-        return $this->version;
     }
 }