+++ src/Services/KeyringManager.php
@@ -2,11 +2,11 @@
 
 namespace AndrewSvirin\Ebics\Services;
 
-use AndrewSvirin\Ebics\Contracts\KeyRingManagerInterface;
-use AndrewSvirin\Ebics\Factories\KeyRingFactory;
+use AndrewSvirin\Ebics\Contracts\KeyringManagerInterface;
+use AndrewSvirin\Ebics\Factories\KeyringFactory;
 
 /**
- * An EbicsKeyRing instance can hold sets of private user keys and/or public
+ * An EbicsKeyring instance can hold sets of private user keys and/or public
  * bank keys. Private user keys are always stored AES encrypted by the
  * specified passphrase (derivated by PBKDF2). For each key file on disk or
  * same key dictionary a singleton instance is created.
@@ -14,15 +14,15 @@ use AndrewSvirin\Ebics\Factories\KeyRingFactory;
  * @license http://www.opensource.org/licenses/mit-license.html  MIT License
  * @author Andrew Svirin
  */
-abstract class KeyRingManager implements KeyRingManagerInterface
+abstract class KeyringManager implements KeyringManagerInterface
 {
-    protected KeyRingFactory $keyRingFactory;
+    protected KeyringFactory $keyringFactory;
 
     /**
      * Constructor.
      */
     public function __construct()
     {
-        $this->keyRingFactory = new KeyRingFactory();
+        $this->keyringFactory = new KeyringFactory();
     }
 }